+++ app/lib/groups/src/pages/course/create/analytics/events/course_create_event.dart
