+++ app/lib/groups/src/pages/course/create/models/course_template.dart
