+++ app/lib/groups/src/pages/course/create/models/user_input.dart
