+++ app/lib/groups/src/pages/course/create/pages/course_template_page.dart
@@ -11,12 +11,12 @@ import 'package:common_domain_models/common_domain_models.dart';
 import 'package:flutter/material.dart';
 import 'package:group_domain_models/group_domain_models.dart';
 import 'package:provider/provider.dart';
-import 'package:sharezone/groups/src/pages/course/create/course_create_page.dart';
-import 'package:sharezone/groups/src/pages/course/create/my_admin_school_classes_provider.dart';
-import 'package:sharezone/groups/src/pages/course/create/src/bloc/course_create_bloc.dart';
-import 'package:sharezone/groups/src/pages/course/create/src/bloc/course_create_bloc_factory.dart';
+import 'package:sharezone/groups/src/pages/course/create/pages/course_create_page.dart';
+import 'package:sharezone/groups/src/pages/course/create/bloc/my_admin_school_classes_provider.dart';
+import 'package:sharezone/groups/src/pages/course/create/bloc/course_create_bloc.dart';
+import 'package:sharezone/groups/src/pages/course/create/bloc/course_create_bloc_factory.dart';
 import 'package:sharezone/groups/src/pages/school_class/card/school_class_card.dart';
-import 'package:sharezone/groups/src/pages/course/create/course_template.dart';
+import 'package:sharezone/groups/src/pages/course/create/models/course_template.dart';
 import 'package:sharezone_common/api_errors.dart';
 import 'package:sharezone_widgets/sharezone_widgets.dart';
