+++ app/lib/main/sharezone_app.dart
@@ -19,7 +19,7 @@ import 'package:sharezone/account/account_page.dart';
 import 'package:sharezone/account/use_account_on_multiple_devices_instruction.dart';
 import 'package:sharezone/blackboard/blackboard_picture.dart';
 import 'package:sharezone/feedback/history/feedback_history_page.dart';
-import 'package:sharezone/groups/src/pages/course/create/course_template_page.dart';
+import 'package:sharezone/groups/src/pages/course/create/pages/course_template_page.dart';
 import 'package:sharezone/main/bloc_dependencies.dart';
 import 'package:sharezone/main/sharezone_bloc_providers.dart';
 import 'package:sharezone/calendrical_events/page/calendrical_events_page.dart';
